test_moves_semidirected.jl working and test_parsimony mostly working.
In the former there was a test for moveTargetUpdate../ This function is in SNaQ and the test got moved to the test_badDiamII.jl in SNaQ

In the latter, the test passes with the exception of a maximum parsimony search. Some of the needed internals were moved to SNaQ, breaking this function. Also the search needs fAbs which got moved to SNaq, I think.